requirements.txt
dj-database-url==0.4.1
Django==1.9.6
django-compat==1.0.11
django-hijack==2.1.1
django-hijack-admin==2.1.1
django-markdown-deux==1.0.5
django-registration-redux==1.4
django-taggit==0.18.3
gunicorn==19.5.0
markdown2==2.3.1
psycopg2==2.6.1
ptyprocess==0.5.1
six==1.10.0
sqlparse==0.1.19
whitenoise==3.1
